refactor: rename core to domain
