+++ src/application/usecases/TokenUseCase.ts
@@ -1,5 +1,5 @@
-import { TokenData } from '../../core/models/TokenData'
-import { ITokenService } from '../../core/services/ITokenService'
+import { TokenData } from '../../domain/models/TokenData'
+import { ITokenService } from '../../domain/services/ITokenService'
 
 /**
  * TokenUseCase
