+++ src/domain/config.ts
