+++ src/domain/interfaces/IAuthRepository.ts
