+++ src/domain/models/CodedError.ts
