+++ src/domain/models/TokenData.ts
