+++ src/domain/services/IAuthService.ts
