+++ src/domain/services/ITokenService.ts
