+++ src/infrastructure/controllers/AuthController.ts
@@ -1,13 +1,13 @@
 import { Request, Response } from 'express'
 import { AuthUseCase } from '../../application/usecases/AuthUseCase'
 import { AuthRepository } from '../adapters/repositories/AuthRepository'
-import { AuthService } from '../../core/services/AuthService'
+import { AuthService } from '../../domain/services/AuthService'
 import { Error } from 'mongoose'
-import { CodedError } from '../../core/models/CodedError'
+import { CodedError } from '../../domain/models/CodedError'
 import { TokenUseCase } from '../../application/usecases/TokenUseCase'
-import { TokenData } from '../../core/models/TokenData'
+import { TokenData } from '../../domain/models/TokenData'
 import { ProfileRepository } from '../adapters/repositories/ProfileRepository'
-import { ProfileService } from '../../core/services/ProfileService'
+import { ProfileService } from '../../domain/services/ProfileService'
 import { ProfileUseCase } from '../../application/usecases/ProfileUseCase'
 import logger from '../Logger'
